teaser image

+++ posts/convert-meta.py
@@ -1,10 +1,12 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
+
 '''
 Script to convert Wordpress tags to canonical format
 '''
 
 import glob
+import os
 
 
 def prettify(tag):
@@ -13,11 +15,24 @@ def prettify(tag):
     return tag
 
 
-for fn in glob.glob('*.meta'):
-    with open(fn) as fd:
-        lines = fd.readlines()
-    words = map(prettify, lines[3].split(','))
-    lines[3] = ','.join(words)
-    print lines
-    with open(fn, 'wb') as fd:
-        fd.writelines(lines)
+# for fn in glob.glob('*.meta'):
+#    with open(fn) as fd:
+#        lines = fd.readlines()
+#    words = map(prettify, lines[3].split(','))
+#    lines[3] = ','.join(words)
+#    print lines
+#    with open(fn, 'wb') as fd:
+#        fd.writelines(lines)
+
+for fn in glob.glob('*.md'):
+    meta_fn = fn[:-2] + 'meta'
+    if os.path.exists(meta_fn):
+        with open(meta_fn) as fd:
+            lines = fd.readlines()
+        with open(fn, 'wb') as fd:
+            oldcontent = fd.read()
+        with open(fn, 'wb') as fd:
+            fd.write('<!--\n')
+            fd.writelines(lines)
+            fd.write('-->\n')
+            fd.write(oldcontent)
